refactor(graph): drop the Value default — graph-core fully generic (VIZ-65)
Moves the `GraphValue` trait and its `impl for Value` into `vizij-api-core`,
next to the value vocabulary they reify. `vizij-graph-core` re-exports the
trait and no longer names `vizij_api_core::Value`: every generic type
(`GraphSpec`, `NodeSpec`, `NodeParams`, `GraphRuntime`, `PortValue`,
`PlanCache`, …) drops its `= Value` default, so the crate is generic over
`GraphValue` with no built-in value binding of its own.

Hosts and consumers now name their value type explicitly (`GraphSpec<Value>`,
etc.) — the point of the change: the concrete value belongs to the host, not
the graph. `WriteOp`/`WriteBatch` keep their `= Value` default (they live in
api-core, where naming Value is fine).
