+++ crates/orchestrator/vizij-orchestrator-core/examples/single_pass.rs
@@ -12,7 +12,7 @@ fn main() -> anyhow::Result<()> {
     // Graph controller
     let gcfg = GraphControllerConfig {
         id: "g".into(),
-        spec: GraphSpec::default().with_cache(),
+        spec: GraphSpec::<vizij_api_core::Value>::default().with_cache(),
         subs: Subscriptions::default(),
     };
     orch = orch.with_graph(gcfg);
