+++ crates/orchestrator/vizij-orchestrator-core/examples/two_pass.rs
@@ -4,7 +4,9 @@ use vizij_api_core::TypedPath;
 use vizij_graph_core::types::GraphSpec;
 use vizij_orchestrator::{GraphControllerConfig, Orchestrator, Schedule, Subscriptions};
 
-fn graph_spec_from_json(mut spec_json: serde_json::Value) -> anyhow::Result<GraphSpec> {
+fn graph_spec_from_json(
+    mut spec_json: serde_json::Value,
+) -> anyhow::Result<GraphSpec<vizij_api_core::Value>> {
     vizij_api_core::json::normalize_graph_spec_value(&mut spec_json)
         .map_err(|e| anyhow::anyhow!(e))?;
     Ok(serde_json::from_value(spec_json)?)
@@ -12,7 +14,7 @@ fn graph_spec_from_json(mut spec_json: serde_json::Value) -> anyhow::Result<Grap
 
 fn main() -> anyhow::Result<()> {
     // Graph A: produce a constant value and expose it via an output node.
-    let producer_spec: GraphSpec = graph_spec_from_json(json!({
+    let producer_spec: GraphSpec<vizij_api_core::Value> = graph_spec_from_json(json!({
         "nodes": [
             {
                 "id": "constant_one",
@@ -31,7 +33,7 @@ fn main() -> anyhow::Result<()> {
     }))?;
 
     // Graph B: consume the shared value, double it, and emit another output.
-    let consumer_spec: GraphSpec = graph_spec_from_json(json!({
+    let consumer_spec: GraphSpec<vizij_api_core::Value> = graph_spec_from_json(json!({
         "nodes": [
             {
                 "id": "shared_input",
